refactor: 쿼리 클라이언트 초기화 로직 개선 및 관련 설정 추가
