+++ apps/web/app/_api/ClientQueryProvider.tsx
@@ -1,51 +1,8 @@
 'use client';
 
-import {
-  defaultShouldDehydrateQuery,
-  QueryClient,
-  QueryClientProvider,
-} from '@tanstack/react-query';
+import { QueryClientProvider } from '@tanstack/react-query';
 
-// https://tkdodo.eu/blog/react-query-as-a-state-manager#customize-staletime
-const STALE_TIME_MS = 20 * 1000; // 20 seconds
-
-function makeQueryClient() {
-  return new QueryClient({
-    defaultOptions: {
-      queries: {
-        retry: 3,
-        refetchOnWindowFocus: false,
-        staleTime: STALE_TIME_MS,
-      },
-      dehydrate: {
-        // include pending queries in dehydration
-        shouldDehydrateQuery: (query) =>
-          defaultShouldDehydrateQuery(query) || query.state.status === 'pending',
-        shouldRedactErrors: () => {
-          // We should not catch Next.js server errors
-          // as that's how Next.js detects dynamic pages
-          // so we cannot redact them.
-          // Next.js also automatically redacts errors for us
-          // with better digests.
-          return false;
-        },
-      },
-    },
-  });
-}
-
-let clientQueryClient: QueryClient | undefined = undefined;
-
-function getQueryClient() {
-  if (typeof window === 'undefined') {
-    // Server: always make a new query client
-    return makeQueryClient();
-  } else {
-    // Browser: make a new query client if we don't already have one
-    if (!clientQueryClient) clientQueryClient = makeQueryClient();
-    return clientQueryClient;
-  }
-}
+import { getQueryClient } from 'app/_util/queryClient';
 
 export const ClientQueryProvider = ({ children }: React.PropsWithChildren) => {
   const queryClient = getQueryClient();
